resolve conditional rendering issue
