+++ mobile/src/contexts/AuthContext.tsx
@@ -10,13 +10,13 @@ const YSC_SERVER_URI = "https://ystemandchess.com/middleware/";
 type AuthState = {
   isLoggedIn: boolean;
   login: (username: string, password: string) => void;
-  validate: () => boolean;
+  validate: () => void;
 };
 
 const initialState: AuthState = {
   isLoggedIn: false,
   login: () => {},
-  validate: () => false,
+  validate: () => {},
 };
 
 export const AuthContext = createContext<AuthState>(initialState);
@@ -25,13 +25,8 @@ export const AuthProvider: React.FC = ({ children }) => {
   const { setUserState } = useContext(UserContext);
   const [isLoggedIn, setIsLoggedIn] = useState(false);
 
-  React.useEffect(() => {
-    console.log("isLoggedIn in auth: ", isLoggedIn);
-  }, [isLoggedIn]);
-
   const login = (username: string, password: string): void => {
     (async () => {
-      try {
         const params = {
           username,
           password,
@@ -40,31 +35,31 @@ export const AuthProvider: React.FC = ({ children }) => {
         const res = await fetch(url, {
           method: "POST",
         });
-        const tokenJson = await res.json();
-        const token = JSON.stringify(tokenJson);
-        // store token & user info
-        await SecureStore.setItemAsync("token", token);
-        const decoded: any = jwt_decode(token);
+        
+        if (res.ok) {
+          const tokenJson = await res.json();
+          const token = JSON.stringify(tokenJson);
+          // store token & user info
+          await SecureStore.setItemAsync("token", token);
+          const decoded: any = jwt_decode(token);
 
-        // store user info in User context
-        const newUserState: User = {
-          username: decoded.username,
-          firstName: decoded.firstName,
-          lastName: decoded.lastName,
-          role: decoded.role,
-          email: decoded.email,
-        };
-        setUserState(newUserState);
-        setIsLoggedIn(true);
-      } catch (err) {
-        // TODO: error catching
-        console.error("invalid login");
-      }
+          // store user info in User context
+          const newUserState: User = {
+            username: decoded.username,
+            firstName: decoded.firstName,
+            lastName: decoded.lastName,
+            role: decoded.role,
+            email: decoded.email,
+          };
+          setUserState(newUserState);
+          setIsLoggedIn(true);
+        } else {
+          console.error("Login request was unsuccessful.")
+        }
     })();
   };
 
-  const validate = (): boolean => {
-    let flag = false;
+  const validate = (): void => {
     (async () => {
       const url = YSC_SERVER_URI + "auth/validate";
       const tokenRes = await SecureStore.getItemAsync("token");
@@ -77,21 +72,19 @@ export const AuthProvider: React.FC = ({ children }) => {
 
       // token is invalid
       if (res.status !== 200) {
-        console.error("couldn't validate token");
-
         // reset user state
         setUserState(initialUser);
         setIsLoggedIn(false);
+
+        console.error("Couldn't validate token.");
       }
 
       // token is valid
       if (res.status === 200) {
-        flag = true;
         setIsLoggedIn(true);
+        console.log("Validated token");
       }
     })();
-    console.log("flag", flag);
-    return flag;
   };
 
   return (
